Enhance opcode to setup SQE in place
Currently the opcode module provides a data structure for each io-uring
operation code. And it works in following way to submit an SQE:
- generate an opcode object, such Nop, Readv etc.
- convert the opcode object into an SQE object
- copy the generated SQE object onto the next available SQE in the
  submission queue

One previous patch in this series implements PrepareSQE to optimize the
process to submit a request as:
- generate an opcode object, such Nop, Readv etc.
- convert the opcode object to an SQE in the submission queue in place

The process to submit a request could be further optimized as:
- get the next available SQE from the submission queue
- prepare the SQE for io uring operation
- commit the SQE

This patch enhance the opcode module to prepare SQE in place, without
involving the intermedia opcode structures. The way to achieve the goal
is to introduce a structure for each opcode to support in-place
preparatio.
    #[repr(transparent)]
    pub struct ReadvSqe {
        sqe: sys::io_uring_sqe,
    }

    impl ReadvSqe {
        #[inline]
        pub fn prepare(&mut self, fd: sealed::Target, iovec: *const libc::iovec, len: u32) {
	}
        #[inline]
        pub fn ioprio(mut self, ioprio: u16) -> Self {
            self.sqe.ioprio = ioprio.as_sqe_value() as _;
            self
        }
    }

    impl<'a> From<&'a mut sys::io_uring_sqe> for &'a mut ReadvSqe {
        #[inline]
        fn from(sqe: &'a mut sys::io_uring_sqe) -> &'a mut ReadvSqe {
            unsafe { mem::transmute(sqe) }
        }
    }

Signed-off-by: Liu Jiang <gerry@linux.alibaba.com>