add register_files_sparse
The ability to register a completely sparse file descriptor table is
introduced with the 5.19 kernel.

Also, a register test module is added with a single test for the new
register_files_sparse submitter method. Besides testing that the method
can be run without error, it tests that unregister_files works by
checking that the register form cannot be done twice in a row and it
checks that the unregister form cannot be done twice in a row.