submit_and_wait: check overflow bit
The uring interface will set the IORING_SQ_CQ_OVERFLOW bit in sq_flags
to indicate some CQE didn't fit in the cqueue and is being held in
overflow memory.

The uring interface will only move entries from the overflow memory to
the cqueue when it receives the 'enter' syscall and the
IORING_ENTER_GETEVENTS bit is set in the flags parameter.

So this condition, of the overflow being set, is a third reason to add
this IORING_ENTER_GETEVENTS  bit to the flags passed to 'enter'.