Make the "unstable" feature have no effect
Dependent crates that rely on "unstable" functionality need to restrict
the set of supported io-uring versions to a predetermined closed range
(e.g. >=0.5.6, <=0.5.8), just in case a future release breaks
compatibility. This makes io-uring version incompatibilities in a
dependency graph more likely, and can complicate packaging for
distributions (see [1] for an example).

Let's rely on the fact that we're pre-1.0 to safely try out new APIs
instead. Still, keep an unused "unstable" feature on the io-uring crate
to avoid breaking compatibility.

[1] https://bugzilla.redhat.com/show_bug.cgi?id=2124697#c76

Signed-off-by: Alberto Faria <afaria@redhat.com>