Add proper remove fail test and disable other one
A funny set of circumstances, if I understand them correctly.

The unit test that was meant to fail because it was originally
designed to use a buffer group id that was not registered

    was now inadvertently using a buffer group id that *was* registered
    because more test cases were added above it, and the next logical
    group id value to use ... was used.

So this unit test to assert an error would be detected, for some
reason still succeeded when run on a 5.15 kernel. Only when running on a
6.1 kernel did the assertion no longer succeed. In looking at it for a
few minutes, it seemed like it shouldn't have succeeded in the 5.15 kernel
either but rather than argue with the results and look for a reason in
kernel diffs between 5.15 and 6.1, I went with it.

Added a new test to assert the expected failure by using a group id that
isn't registered, and commented out the case that fails on 5.15 and
works on 6.1 in case anyone else wants to investigate further.