test/register-buf-ring: provide non-zero length argument to read op (#173)
This tweaks the `register_buf_ring` test, now passing a non-zero read length argument to the I/O operation.

Background:

The io_uring device can perform `read` operations into buffers that it has been given previously, namely through `provided buffers` or one or more `buffer rings`. As a result, the device knows the length of the buffers before the `read` operation is even submitted.

For kernel 5.19, passing a length value of zero was treated as nothing special - the user asked for zero bytes to be read so no bytes were read.

Starting with kernel 6.0, the length value of zero is interpreted by the device to mean the user intends for the full length of the buffer that is chosen to be used as the `read` length value.