ci: use host kernel for tests
Until we find an ergonomic way to virtualize aarch64 within github
actions, let's stick with only running the test package on x86. Unit
tests will run on both architectures via cross.

This bumps the kernel version used for tests from 5.10 to 5.15. To run
newer kernel versions (and to reintroduce running the test package on
aarch64) we'll need to revisit virtualization with QEMU.